# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CTF Full V1 API
  version: 0.1.0
  description: >
    Partial contract skeleton for VS01-VS05 and the read-only External Reality
    Interface. All paths below are relative to the /api/v1 server URL. The runnable
    FastAPI application's generated /openapi.json is authoritative where this
    skeleton does not yet declare an implemented endpoint or exact response shape.
    Domain implementations must expand schemas without weakening ownership, version,
    evidence, genealogy, idempotency, or human-authority requirements. Every
    consequential mutation requires Idempotency-Key in the deployed policy and
    may return 413 or 429 with a stable error envelope.
...
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /projects:
    post:
      tags: [Projects]
      operationId: createProject
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateProject"}
      responses:
        "201": {$ref: "#/components/responses/ProjectResponse"}
        "409": {$ref: "#/components/responses/Conflict"}
...
  /projects/{projectId}/evidence/analyze-document:
    parameters: [{$ref: "#/components/parameters/ProjectId"}]
    post:
      tags: [Document Intelligence]
      operationId: analyzeDocumentOffline
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}]
      description: >
        Creates or returns a deduplicated durable local job. Document text is
        treated only as untrusted text; no LLM is invoked and all extracted
        claims/evidence remain unconfirmed candidates.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/DocumentAnalysisRequest"}
      responses:
        "202":
          description: Durable queued or existing document job
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DocumentJob"}
        "404": {$ref: "#/components/responses/NotFound"}
  /projects/{projectId}/document-jobs/{jobId}:
    parameters:
      - {$ref: "#/components/parameters/ProjectId"}
      - {name: jobId, in: path, required: true, schema: {type: string}}
    get:
      tags: [Document Intelligence]
      operationId: getDocumentJob
      responses:
        "200":
          description: Authorized job status, progress, counts and safe error
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DocumentJob"}
        "404": {$ref: "#/components/responses/NotFound"}
  /projects/{projectId}/attachments/{attachmentId}/parsed:
    parameters:
      - {$ref: "#/components/parameters/ProjectId"}
      - {name: attachmentId, in: path, required: true, schema: {type: string}}
      - {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}}
      - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 100}}
    get:
      tags: [Document Intelligence]
      operationId: getParsedDocument
      responses:
        "200":
          description: Parsed metadata and paginated provenance-bearing chunks
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ParsedDocumentResponse"}
        "404": {$ref: "#/components/responses/NotFound"}
...
  /projects/{projectId}/resources/{kind}/{resourceId}/supersede:
    parameters:
      - {$ref: "#/components/parameters/ProjectId"}
      - {name: kind, in: path, required: true, schema: {type: string}}
      - {name: resourceId, in: path, required: true, schema: {type: string}}
    post:
      tags: [Projects]
      operationId: supersedeConfirmedResource
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}]
      description: Preserves the immutable confirmed record and creates a linked replacement.
      requestBody:
        required: true
        content: {application/json: {schema: {type: object}}}
      responses:
        "201": {description: New replacement with supersedes_id}
        "409": {$ref: "#/components/responses/Conflict"}
...
  /projects/{projectId}/creation-graph:
    parameters: [{$ref: "#/components/parameters/ProjectId"}]
    get:
      tags: [Creation]
      operationId: getCreationGraph
      responses:
        "200":
          description: Persisted genealogy only
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CreationGraph"}
...
  /projects/{projectId}/execution-events:
    parameters: [{$ref: "#/components/parameters/ProjectId"}]
    post:
      tags: [Execution]
      operationId: createExecutionEvent
      parameters: [{$ref: "#/components/parameters/IdempotencyKey"}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ExecutionEvent"}
      responses:
        "201":
          description: Event with deterministic minimum materiality applied
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ExecutionEvent"}
...
components:
  securitySchemes:
    bearerAuth: {type: apiKey, in: header, name: X-Session-Token}
    providerCredential: {type: apiKey, in: header, name: X-Provider-Credential}
  parameters:
    ProjectId: {name: projectId, in: path, required: true, schema: {type: string, format: uuid}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 1, maxLength: 200}}
...
  responses:
    ProjectResponse:
      description: Project
      content: {application/json: {schema: {$ref: "#/components/schemas/Project"}}}
...
    NotFound: {$ref: "#/components/responses/ErrorResponse"}
    Conflict: {$ref: "#/components/responses/ErrorResponse"}
...
    ErrorResponse:
      description: Stable safe error
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
  schemas:
...
    CreateProject:
      type: object
      required: [entry_family, initial_input]
      properties:
        entry_family: {enum: [CREATION, FUNDING, DOCUMENT]}
        entry_type: {type: string}
        initial_input: {type: string, minLength: 1}
        source: {type: object, additionalProperties: {type: string}}
    Project:
      type: object
      required: [id, stage, version, created_at]
      properties:
        id: {type: string, format: uuid}
        stage: {type: string}
        version: {type: integer, minimum: 1}
        created_at: {type: string, format: date-time}
...
    DocumentAnalysisRequest:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [attachment_id]
          properties:
            attachment_id: {type: string}
    DocumentJob:
      type: object
      required: [id, project_id, kind, status, data]
      properties:
        id: {type: string}
        project_id: {type: string}
        kind: {const: DOCUMENT_JOB}
        status: {enum: [QUEUED, PROCESSING, COMPLETED, FAILED]}
        data:
          type: object
          required: [attachment_id, status, progress, counts]
          properties:
            attachment_id: {type: string}
            attachment_checksum_sha256: {type: string}
            status: {enum: [QUEUED, PROCESSING, COMPLETED, FAILED]}
            progress: {type: integer, minimum: 0, maximum: 100}
            counts: {type: object, additionalProperties: {type: integer, minimum: 0}}
            error:
              oneOf:
                - type: object
                  required: [code, message]
                  properties:
                    code: {type: string}
                    message: {type: string}
                - type: "null"
    DocumentChunk:
      type: object
      required: [id, kind, status, data]
      properties:
        id: {type: string}
        kind: {const: DOCUMENT_CHUNK}
        status: {const: PARSED}
        data:
          type: object
          required: [attachment_id, checksum_sha256, ordinal, text, provenance]
          properties:
            attachment_id: {type: string}
            checksum_sha256: {type: string}
            ordinal: {type: integer, minimum: 1}
            text: {type: string, maxLength: 4000}
            provenance:
              type: object
              properties:
                page: {type: integer, minimum: 1}
                sheet: {type: string}
                section: {type: string}
                row_start: {type: integer, minimum: 1}
                row_end: {type: integer, minimum: 1}
    ParsedDocumentResponse:
      type: object
      required: [document, chunks, pagination]
      properties:
        document: {type: object}
        chunks: {type: array, items: {$ref: "#/components/schemas/DocumentChunk"}}
        pagination:
          type: object
          required: [offset, limit, total]
          properties:
            offset: {type: integer, minimum: 0}
            limit: {type: integer, minimum: 1, maximum: 200}
            total: {type: integer, minimum: 0}
    CreationGraph:
      type: object
      required: [nodes, links]
      properties:
        nodes: {type: array, items: {type: object}}
        links:
          type: array
          items:
            type: object
            required: [from_type, from_id, to_type, to_id, relation]
            properties:
              from_type: {type: string}
              from_id: {type: string}
              to_type: {type: string}
              to_id: {type: string}
              relation: {type: string}
...
    ExecutionEvent:
      type: object
      required: [type, occurred_at, materiality]
      properties:
        type: {type: string}
        occurred_at: {type: string, format: date-time}
        materiality: {enum: [LOCAL, MATERIAL, DECISION_RELEVANT]}
        evidence_refs: {type: array, items: {type: string}}
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: {type: string}
            message: {type: string}
            request_id: {type: string}